+++ db/migrate/20250120221639_create_plans.rb
@@ -7,10 +7,8 @@ class CreatePlans < ActiveRecord::Migration[8.0]
       t.string :name, null: false, default: ""
       t.integer :max_users, null: false, default: 0
       t.integer :max_companies, null: false, default: 0
-      t.decimal :free_price, precision: 10, scale: 2, null: false, default: 0.00
-      t.decimal :monthly_price, precision: 10, scale: 2, null: false, default: 0.00
-      t.decimal :semi_annual_price, precision: 10, scale: 2, null: false, default: 0.00
-      t.decimal :annual_price, precision: 10, scale: 2, null: false, default: 0.00
+      t.integer :plan_type, null: false, default: 0
+      t.decimal :plan_price, precision: 10, scale: 2, null: false, default: 0.00
       t.integer :status, null: false, default: 0
       t.datetime :deleted_at
 
@@ -20,10 +18,8 @@ class CreatePlans < ActiveRecord::Migration[8.0]
     add_index :plans, :name
     add_index :plans, :max_users
     add_index :plans, :max_companies
-    add_index :plans, :free_price
-    add_index :plans, :monthly_price
-    add_index :plans, :semi_annual_price
-    add_index :plans, :annual_price
+    add_index :plans, :plan_type
+    add_index :plans, :plan_price
     add_index :plans, :status
     add_index :plans, :deleted_at
   end