+++ spec/factories/plans.rb
@@ -5,9 +5,8 @@ FactoryBot.define do
     name { Faker::Company.buzzword }
     max_users { rand(1..99) }
     max_companies { rand(1..99) }
-    free_price { rand(1..999) }
-    monthly_price { rand(1..999) }
-    semi_annual_price { rand(1..999) }
-    annual_price { rand(1..999) }
+    plan_price { rand(1..999) }
+    plan_type { [ :free, :monthly, :semi_annual, :annual ].sample }
+    status { :active }
   end
 end